Simplify the freeing of the result list
Flagged by PVS Studio:

V809. Verifying that a pointer value is not NULL is not
required. The 'if (ptr != NULL)' check can be removed.

The analyzer has detected a code fragment that can be
simplified. The 'free()' function and 'delete' operator handle
the null pointer correctly. So we can remove the pointer
check. This allows us to delete an unnecessary string to make
the code shorter and clearer.