Replace memset with {0}
Initializing to 0 with {0} is consider a better practice. Safe and allows the compiler to perform the best optimization it can.

Right now I'm on Window and forego the Linux part of codes because I wish to ensure it work. There appears to be no issue at all with this change here.