Eliminating Pedantic warnings
I am on MinGW GCC 13.2 So with -Wpedantic you'll see a bunch of warning pops up in ENet.
It involved the mixing of two separate enums with their | operations. We resolve this by using (int) which should generate absolutely no extra codes and remove the annoying messages.

There was also a mismatch of types in one of the printf when using ENET_DEBUG. So I corrected that at line: 3185-92

Finally the use of ENET_ATOMIC_CAS in enet_get_time() with the implementation of line 1222-28 ISO forbid the use of ({...)} but here where it's implemented it only occurs for GCC which make the expression safe and valid. To ignore such message we would have to use at line: 5101
#if defined(__GNUC__) // Ignore warning.
#pragma GCC diagnostic push
#pragma GCC diagnostic ignored "-Wpedantic"
#endif
uint64_t old_value = ENET_ATOMIC_CAS(&start_time_ns, 0, want_value);
#if defined(__GNUC__)
#pragma GCC diagnostic pop
#endif