Should we be nullifying the pointer after it's free?
I trust that this software is reliable. But I do see that pointers aren't being nullified here. So it's a bit of the headscratcher for me. I'll leave this to the judgment of the author.

I like performance, so having less code is always a plus. But I also value security.