Supporting dependency injection via direct use of external socket.
I added two things. a function is_udp_socket(socket) to ensure it's UDP. and a bool ownsSocket for host to ensure it behave with internal mechanism for socket or to use external without creating or closing socket or changing the configuration of it. (Leaving it all to the implementer of their own network.)

You use your own socket by putting it in the new argument of enet_host_create() ENetSocket* externalUDPSocket.

You set it to NULL to use the internal socket as normally done.

This should theoretically work. I am unable to test it for myself because I don't need to make my own network so I'm not sure how to do it either.

Things to test:
Use of external socket should work.
is_udp_socket should be valid.
External socket should only be managed by the programmer.