Supporting dependency injection. Re-request
I needed to re-request #71 because codespace wasn't working for me.

The change I made is an attempt to address the API-compatibility concern. I assumed it had to do with how adding an extra parameter to existing function (enet_host_create) would cause inconsistence requirement. So I addressed this by moving it over to enet_host_create_ext() and reuse the code with enet_host_create calling it set to NULL. Making it behave exactly as one would expect.

Hope this is a proper solution! :)